Avoid circular dependencies and import un-bundled rrdom
